Better encapsulation of JavaScript code test
Embeding the code into an HTML element causes trouble is the JavaScript
code contains HTML characters.  There were also problems with wrapping
the code into a multiline string, because if the code contains back-tick
characters and placeholders that causes trouble too.

In the end, I concluded the best is to split the text into several lines
and concatenate them.  Prior to that is necessary to escape
doule-quotes.  This seems to work nicely in all cases.