Handle 'SyntaxError' tests
There are tests that emits SyntaxError. Such tests have a 'type' flag
which value is 'SyntaxError'.  This means it's necessary to pass the
whole test attributes array to the iframe's test and in case it's a test
of this type wrap the code in a try/catch while evaluating the code with
eval.